move Queues to its own mod, rename to SingleThreadQueues
this should hopefully make it clearer that these are not used by the
threadpool.
