+++ tokio/src/task/local.rs
@@ -1,7 +1,6 @@
 //! Runs `!Send` futures on the current thread.
-use crate::runtime::basic_scheduler::Queues;
 use crate::sync::AtomicWaker;
-use crate::task::{self, JoinHandle, Schedule, Task};
+use crate::task::{self, queue::SingleThreadQueues, JoinHandle, Schedule, Task};
 
 use std::cell::Cell;
 use std::future::Future;
@@ -85,7 +84,7 @@ cfg_rt_util! {
 struct Scheduler {
     tick: Cell<u8>,
 
-    queues: Queues<Self>,
+    queues: SingleThreadQueues<Self>,
 
     /// Used to notify the `LocalFuture` when a task in the local task set is
     /// notified.
@@ -354,7 +353,7 @@ impl Scheduler {
     fn new() -> Self {
         Self {
             tick: Cell::new(0),
-            queues: Queues::new(),
+            queues: SingleThreadQueues::new(),
             waker: AtomicWaker::new(),
         }
     }
